Changed so surfstart is generic and points back

+++ configure.sh
@@ -2,9 +2,10 @@
 echo "SurfingPi Configure"
 echo
 read -n1 -p "Do you want to update surfshark configs first? [y,n]  " doit
-case $doit in  
-  y|Y) echo "Updating configs"  
-  	cd /etc/openvpn
+case $doit in
+y | Y)
+      echo "Updating configs"
+      cd /etc/openvpn
       sudo rm *
       sudo wget https://my.surfshark.com/vpn/api/v1/server/configurations
       sudo unzip configurations
@@ -13,9 +14,10 @@ case $doit in
 
       sudo grep -rl "auth-user-pass" /etc/openvpn/ | sudo xargs sed -i "s>auth-user-pass>auth-user-pass /home/$USER/.surfshark/surf>g"
       echo "done"
-      echo ;;
-  n|N) echo "Skipping update" ;; 
-  *) echo "Please answer y or n" ;; 
+      echo
+      ;;
+n | N) echo "Skipping update" ;;
+*) echo "Please answer y or n" ;;
 esac
 echo
 echo "Surfshark Credentials"
@@ -27,66 +29,60 @@ echo
 read -p "username:" user_var
 read -p "password:" pass_var
 
-if [ -z "$user_var" ]
-then
+if [ -z "$user_var" ]; then
       echo "Leave user and pass the same"
 else
       echo "Updating user and pass"
       sudo rm /home/$USER/.surfshark/surf
       mkdir /home/$USER/.surfshark/
-      sudo echo $user_var > /home/$USER/.surfshark/surf
-      sudo echo $pass_var >> /home/$USER/.surfshark/surf
+      sudo echo $user_var >/home/$USER/.surfshark/surf
+      sudo echo $pass_var >>/home/$USER/.surfshark/surf
       sudo chmod -R 600 /home/$USER/.surfshark
 fi
 
-echo 
-echo "Select which protocol to use"
-select proto in TCP UDP
-do
-echo "$proto selected"
-
-case $proto in
-	TCP)
-tcpendpoints=$(ls /etc/openvpn/ | grep tcp)
-select tcpendpoint in $tcpendpoints
-do
-echo "$tcpendpoint endpoint selected"
-sudo chmod -R 777 /opt/SurfingPi
-sudo echo "#!/bin/bash" > /opt/SurfingPi/surfstart
-sudo echo "sudo openvpn /etc/openvpn/$tcpendpoint" >> /opt/SurfingPi/surfstart
-echo "Adding to /usr/bin/"
 echo
-sudo cp /opt/SurfingPi/surfstart /usr/bin/surfstart
-sudo rm /opt/SurfingPi/surfstart
+sudo cp /opt/SurfingPi/surfpointer.sh /usr/bin/surfstart
 sudo chmod +x /usr/bin/surfstart
-echo "surfstart created!"
-sudo chmod -R 755 /opt/SurfingPi
-break
-done
 
-break
-;;
-UDP)
-udpendpoints=$(ls /etc/openvpn/ | grep udp)
-select udpendpoint in $udpendpoints
-do
-echo "$udpendpoint endpoint selected"
-sudo chmod -R 777 /opt/SurfingPi
-echo "#!/bin/bash" > /opt/SurfingPi/surfstart
-echo "sudo openvpn /etc/openvpn/$udpendpoint" >> /opt/SurfingPi/surfstart
-echo "Adding to /usr/bin"
-echo
-sudo cp /opt/SurfingPi/surfstart /usr/bin/surfstart
-sudo rm /opt/SurfingPi/surfstart
-sudo /opt/SurfingPi/chmod +x /usr/bin/surfstart
-echo "surfstart created!"
-sudo chmod -R 755 /opt/SurfingPi
-break
-done
+echo "Select which protocol to use"
+select proto in TCP UDP; do
+      echo "$proto selected"
 
-break
-;;
-esac
-break
-done
+      case $proto in
+      TCP)
+            tcpendpoints=$(ls /etc/openvpn/ | grep tcp)
+            select tcpendpoint in $tcpendpoints; do
+                  echo "$tcpendpoint endpoint selected"
+                  sudo chmod -R 777 /opt/SurfingPi
+                  sudo echo "#!/bin/bash" >/opt/SurfingPi/surfstart
+                  sudo echo "sudo openvpn /etc/openvpn/$tcpendpoint" >>/opt/SurfingPi/surfstart
+                  echo "Adding to /usr/bin/"
+                  echo
+                  sudo chmod +x /opt/SurfingPi/surfstart
+                  echo "surfstart created!"
+                  sudo chmod -R 755 /opt/SurfingPi
+                  break
+            done
 
+            break
+            ;;
+      UDP)
+            udpendpoints=$(ls /etc/openvpn/ | grep udp)
+            select udpendpoint in $udpendpoints; do
+                  echo "$udpendpoint endpoint selected"
+                  sudo chmod -R 777 /opt/SurfingPi
+                  echo "#!/bin/bash" >/opt/SurfingPi/surfstart
+                  echo "sudo openvpn /etc/openvpn/$udpendpoint" >>/opt/SurfingPi/surfstart
+                  echo "Adding to /usr/bin"
+                  echo
+                  sudo chmod +x /opt/SurfingPi/surfstart
+                  echo "surfstart created!"
+                  sudo chmod -R 755 /opt/SurfingPi
+                  break
+            done
+
+            break
+            ;;
+      esac
+      break
+done
